Move codegenerator back to public namespace
Fixes #1199
