+++ codegenerator/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
         "doc.go",
         "parse_req.go",
     ],
-    importpath = "github.com/grpc-ecosystem/grpc-gateway/internal/codegenerator",
+    importpath = "github.com/grpc-ecosystem/grpc-gateway/codegenerator",
     deps = [
         "@com_github_golang_protobuf//proto:go_default_library",
         "@io_bazel_rules_go//proto/wkt:compiler_plugin_go_proto",
