+++ codegenerator/parse_req_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/golang/protobuf/proto"
 	plugin "github.com/golang/protobuf/protoc-gen-go/plugin"
-	"github.com/grpc-ecosystem/grpc-gateway/internal/codegenerator"
+	"github.com/grpc-ecosystem/grpc-gateway/codegenerator"
 )
 
 var parseReqTests = []struct {
